ref: some reorganization of files
- move `class Config` in its own `config.py`
- mov workflow.yaml one level up becuase the `config` dir becomes
  confusing with `config.py`. Also not really needed to have that
  folder with only this file in it.
